+++ tests/CasterToNumberTest.php
@@ -151,6 +151,7 @@ final class CasterToNumberTest extends TestCase
     public function testNullThrows(): void
     {
         $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessageIs('Cannot convert null to Number');
         Caster::toNumber(null);
     }